// src/app/model/org.ts
export class Org{
    id: Number;
    orgName: String;
    orgBusinessAddress: String;
    orgCommAddress: String;
    orgCity: String;
    orgState: String;
    orgCountry: String;
    orgZipcode: String;
    orgPhone: String;
    orgAltPhone: String;
    orgEmail: String;
    orgContactName: String;
    orgContactPhone: String;
    orgContactEmail: String;
    orgBuildVersion: String;
    orgLangPref: String;
    orgTimeZone: String;
    org2fa: Boolean;
    org2faemail: Boolean;
    org2fasms: Boolean;
    orgMode: String;
    orgStatus: boolean;
}